Changed getCurrentLocation to work with async transitions
Before the wrong location would be used in scroll-behavior resulting in
that the wrong scroll state was saved for async transitions, when the
actual transition happens after something async has happened.

This change makes sure we only change the location when we have made the
actual transition.

This change is backwards compatible and will work when using in as
before since we will update the location immediately as we get new
props.